FEAT: código das principais tabelas do sistema e reestruturação das models
- Melhorias na logServices
- Connection com BD é uma service agora
- Error handler
- Reestruturação das Models para automatizar código
- Controllers com automatização
- Rotas em arquivos separados
- Crud de: usuários, adm, obras, portifólio e tags

+++ src/Backend/models/Model.js
@@ -1,5 +1,5 @@
-const Connection = require('../Connection')
-const LogService = require('../services/LogService')
+const ConnectionService = require('../services/ConnectionService');
+const APIError = require('../services/ErrorService');
 
 class Model {
     tableName; // Nome da tabela no banco de dados
@@ -10,40 +10,46 @@ class Model {
         this.identifierColumns = params['identifierColumns'];
     }
 
+    // Verifica se todas as colunas de identificação estão setadas
+    verifyIdentifierColumns(){
+        const identifierParams = this.getIdentifierColumnsObject();
+
+        let identifierParamsSetted = true;
+
+        Object.values(identifierParams).forEach(element => {
+            if (!element) {
+                identifierParamsSetted = false
+            }
+        })
+
+        return identifierParamsSetted;
+    }
+
     // Retorna como Entries ([[key, value]]) todos os atributos da classe que não sejam de lógica como tableName e identifierColumns
-    getRealModelParamsEntries() {
+    getObjectEntries() {
         return Object.entries(this).filter(entrie => {
             return entrie[0] != 'tableName' && entrie[0] != 'identifierColumns'
         })
     }
 
-    getRealModelObject() {
+
+    // Retorna como objeto todos os atributos da classe que não sejam como tableName e identifierColumns
+    getObject() {
         const object = {}
 
-        this.getRealModelParamsEntries().forEach(entrie => {
+        this.getObjectEntries().forEach(entrie => {
             object[entrie[0]] = entrie[1]
         })
 
         return object;
     }
 
-    // Retorna como um objeto todos os atributos reais da classe da forma que devem ser passados como parâmetro para queries 
-    getModelAsSQLParams() {
-        const object = {};
-
-        this.getRealModelParamsEntries().forEach(entrie => {
-            object[`$${entrie[0]}`] = entrie[1]
-        })
-
-        return object;
-    }
-
     // Retorna como um objeto todos os atributos de identificação da classe da forma que devem ser passados como parâmetro para queries 
-    getModelIdentifierColumnsAsSQLParams() {
+    getIdentifierColumnsObject() {
         const object = {};
 
-        this.getRealModelParamsEntries().filter(entrie => this.identifierColumns.includes(entrie[0])).forEach(entrie => {
-            object[`$${entrie[0]}`] = entrie[1]
+        this.getObjectEntries().filter(entrie => this.identifierColumns.includes(entrie[0])).forEach(entrie => {
+            object[`${entrie[0]}`] = entrie[1]
         })
 
         return object
@@ -51,72 +57,87 @@ class Model {
 
     // Seta os atributos da classe de acordo com o array ou objeto passado por parâmetro dinâmicamente
     async setRealModelAttributes(attributes) {
-        return this.getRealModelParamsEntries().forEach(entrie => {
+        return this.getObjectEntries().forEach(entrie => {
             this[entrie[0]] = attributes[entrie[0]]
         })
     }
 
     // Obtem uma tupla da tabela de acordo com as colunas identificadoras
     async getByIdentifierColumns() {
-        return Connection.get(
+        return ConnectionService.get(
             `SELECT ${this.tableName}.* 
             FROM ${this.tableName} 
             WHERE ${this.identifierColumns.map(key => `${key} = $${key}`).join(" AND ")}`,
-            this.getModelIdentifierColumnsAsSQLParams()
+            this.getIdentifierColumnsObject()
         )
     }
 
     static async getSQL(query, params) {
-        return Connection.get(query, params);
+        return ConnectionService.get(query, params);
     }
 
     static async allSQL(query, params) {
-        return Connection.all(query, params);
+        return ConnectionService.all(query, params);
     }
 
     // Insere um registro no banco de dados com os valores que estão na classe
     async insert() {
-        return Connection.insert(
-            `INSERT INTO ${this.tableName} 
-                (${this.getRealModelParamsEntries().map(entrie => entrie[0]).join(", ")})
-            VALUES
-                (${this.getRealModelParamsEntries().map(entrie => `$${entrie[0]}`).join(", ")})`,
-            this.getModelAsSQLParams()
-        )
+        const query = 
+        `INSERT INTO ${this.tableName} 
+            (${this.getObjectEntries().map(entrie => entrie[0]).join(", ")})
+        VALUES
+            (${this.getObjectEntries().map(entrie => `$${entrie[0]}`).join(", ")})`;
+
+        if (!this.verifyIdentifierColumns()) {
+            const id = await ConnectionService.insert(
+                query,
+                this.getObject()
+            )
+
+            if(this.identifierColumns.length == 1){
+                this[this.identifierColumns[0]] = id;
+            }
+
+            return id;
+        } else {
+            throw new APIError(`Objeto não pode ser criado se já possui um identificador`, 403, query)
+        }
     }
 
     static async insertSQL(query, params) {
-        return Connection.insert(query, params);
+        return ConnectionService.insert(query, params);
     }
 
     // Atualiza no banco de dados a tupla referente ao objeto atual
     async update() {
-        if (this.identifierColumns.length) { // Só aplica o update se existirem colunas de identificação setadas, previnindo atualizações em toda a tabela por acidente
-            return Connection.update(
+        if (this.verifyIdentifierColumns() && this.identifierColumns.length) { // Só aplica o update se as colunas identificadoras estiverem setadas
+            return ConnectionService.update(
                 `UPDATE ${this.tableName} 
-                    SET ${this.getRealModelParamsEntries().map(entrie => `${entrie[0]} = $${entrie[0]}`).join(", ")} 
+                    SET ${this.getObjectEntries().map(entrie => `${entrie[0]} = $${entrie[0]}`).join(", ")} 
                 WHERE ${this.identifierColumns.map(key => `${key} = $${key}`).join(" AND ")}`,
-                this.getModelAsSQLParams()
+                this.getObject()
             )
+        } else {
+            throw new APIError(`Objeto não pode ser editado pois não existe um identificador`, 403, query)
         }
     }
 
     static async updateSQL(query, params) {
-        return Connection.update(query, params);
+        return ConnectionService.update(query, params);
     }
 
     // Deleta um registro no banco de dados referente ao objeto atual
     async delete() {
-        if (this.identifierColumns.length) { // Só aplica o delete se existirem colunas de identificação setadas, previnindo exclusões em toda a tabela por acidente
-            return Connection.delete(
+        if (this.verifyIdentifierColumns() && this.identifierColumns.length) { // Só aplica o delete se existirem colunas de identificação setadas, previnindo exclusões em toda a tabela por acidente
+            return ConnectionService.delete(
                 `DELETE FROM ${this.tableName} WHERE ${this.identifierColumns.map(key => `${key} = $${key}`).join(" AND ")}`,
-                this.getModelIdentifierColumnsAsSQLParams()
+                this.getIdentifierColumnsObject()
             )
         }
     }
 
     static async deleteSQL(query, params) {
-        return Connection.delete(query, params);
+        return ConnectionService.delete(query, params);
     }
 }
